[SPARK-22461][ML] Move Spark ML model summaries into a dedicated package
