+++ mllib/src/main/scala/org/apache/spark/ml/regression/GeneralizedLinearRegression.scala
@@ -20,29 +20,29 @@ package org.apache.spark.ml.regression
 import java.util.Locale
 
 import breeze.stats.{distributions => dist}
-import org.apache.commons.lang3.StringUtils
 import org.apache.hadoop.fs.Path
 
 import org.apache.spark.SparkException
 import org.apache.spark.annotation.{Experimental, Since}
 import org.apache.spark.internal.Logging
 import org.apache.spark.ml.PredictorParams
-import org.apache.spark.ml.attribute.AttributeGroup
 import org.apache.spark.ml.feature.{Instance, OffsetInstance}
-import org.apache.spark.ml.linalg.{BLAS, Vector, Vectors}
+import org.apache.spark.ml.linalg.{BLAS, Vector}
 import org.apache.spark.ml.optim._
 import org.apache.spark.ml.param._
 import org.apache.spark.ml.param.shared._
+import org.apache.spark.ml.summary.GeneralizedLinearRegressionSummary
+import org.apache.spark.ml.summary.GeneralizedLinearRegressionTrainingSummary
 import org.apache.spark.ml.util._
 import org.apache.spark.rdd.RDD
-import org.apache.spark.sql.{Column, DataFrame, Dataset, Row}
+import org.apache.spark.sql.{DataFrame, Dataset, Row}
 import org.apache.spark.sql.functions._
 import org.apache.spark.sql.types.{DataType, DoubleType, StructType}
 
 /**
  * Params for Generalized Linear Regression.
  */
-private[regression] trait GeneralizedLinearRegressionBase extends PredictorParams
+private[ml] trait GeneralizedLinearRegressionBase extends PredictorParams
   with HasFitIntercept with HasMaxIter with HasTol with HasRegParam with HasWeightCol
   with HasSolver with Logging {
 
@@ -154,11 +154,11 @@ private[regression] trait GeneralizedLinearRegressionBase extends PredictorParam
   def getOffsetCol: String = $(offsetCol)
 
   /** Checks whether weight column is set and nonempty. */
-  private[regression] def hasWeightCol: Boolean =
+  private[ml] def hasWeightCol: Boolean =
     isSet(weightCol) && $(weightCol).nonEmpty
 
   /** Checks whether offset column is set and nonempty. */
-  private[regression] def hasOffsetCol: Boolean =
+  private[ml] def hasOffsetCol: Boolean =
     isSet(offsetCol) && $(offsetCol).nonEmpty
 
   /** Checks whether we should output link prediction. */
@@ -474,7 +474,7 @@ object GeneralizedLinearRegression extends DefaultParamsReadable[GeneralizedLine
   /**
    * Wrapper of family and link combination used in the model.
    */
-  private[regression] class FamilyAndLink(val family: Family, val link: Link) extends Serializable {
+  private[ml] class FamilyAndLink(val family: Family, val link: Link) extends Serializable {
 
     /** Linear predictor based on given mu. */
     def predict(mu: Double): Double = link.link(family.project(mu))
@@ -516,7 +516,7 @@ object GeneralizedLinearRegression extends DefaultParamsReadable[GeneralizedLine
     }
   }
 
-  private[regression] object FamilyAndLink {
+  private[ml] object FamilyAndLink {
 
     /**
      * Constructs the FamilyAndLink object from a parameter map
@@ -541,7 +541,7 @@ object GeneralizedLinearRegression extends DefaultParamsReadable[GeneralizedLine
    *
    * @param name the name of the family.
    */
-  private[regression] abstract class Family(val name: String) extends Serializable {
+  private[ml] abstract class Family(val name: String) extends Serializable {
 
     /** The default link instance of this family. */
     val defaultLink: Link
@@ -674,7 +674,7 @@ object GeneralizedLinearRegression extends DefaultParamsReadable[GeneralizedLine
    * Gaussian exponential family distribution.
    * The default link for the Gaussian family is the identity link.
    */
-  private[regression] object Gaussian extends Tweedie(0.0) {
+  private[ml] object Gaussian extends Tweedie(0.0) {
 
     override val name: String = "gaussian"
 
@@ -712,7 +712,7 @@ object GeneralizedLinearRegression extends DefaultParamsReadable[GeneralizedLine
    * Binomial exponential family distribution.
    * The default link for the Binomial family is the logit link.
    */
-  private[regression] object Binomial extends Family("binomial") {
+  private[ml] object Binomial extends Family("binomial") {
 
     val defaultLink: Link = Logit
 
@@ -764,7 +764,7 @@ object GeneralizedLinearRegression extends DefaultParamsReadable[GeneralizedLine
    * Poisson exponential family distribution.
    * The default link for the Poisson family is the log link.
    */
-  private[regression] object Poisson extends Tweedie(1.0) {
+  private[ml] object Poisson extends Tweedie(1.0) {
 
     override val name: String = "poisson"
 
@@ -801,7 +801,7 @@ object GeneralizedLinearRegression extends DefaultParamsReadable[GeneralizedLine
    * Gamma exponential family distribution.
    * The default link for the Gamma family is the inverse link.
    */
-  private[regression] object Gamma extends Tweedie(2.0) {
+  private[ml] object Gamma extends Tweedie(2.0) {
 
     override val name: String = "gamma"
 
@@ -838,7 +838,7 @@ object GeneralizedLinearRegression extends DefaultParamsReadable[GeneralizedLine
    *
    * @param name the name of link function.
    */
-  private[regression] abstract class Link(val name: String) extends Serializable {
+  private[ml] abstract class Link(val name: String) extends Serializable {
 
     /** The link function. */
     def link(mu: Double): Double
@@ -850,7 +850,7 @@ object GeneralizedLinearRegression extends DefaultParamsReadable[GeneralizedLine
     def unlink(eta: Double): Double
   }
 
-  private[regression] object Link {
+  private[ml] object Link {
 
     /**
      * Gets the [[Link]] object based on param family, link and linkPower.
@@ -883,7 +883,7 @@ object GeneralizedLinearRegression extends DefaultParamsReadable[GeneralizedLine
   }
 
   /** Power link function class */
-  private[regression] class Power(val linkPower: Double)
+  private[ml] class Power(val linkPower: Double)
     extends Link("power") {
 
     override def link(mu: Double): Double = {
@@ -911,7 +911,7 @@ object GeneralizedLinearRegression extends DefaultParamsReadable[GeneralizedLine
     }
   }
 
-  private[regression] object Identity extends Power(1.0) {
+  private[ml] object Identity extends Power(1.0) {
 
     override val name: String = "identity"
 
@@ -1148,423 +1148,6 @@ object GeneralizedLinearRegressionModel extends MLReadable[GeneralizedLinearRegr
 
       DefaultParamsReader.getAndSetParams(model, metadata)
       model
-    }
-  }
-}
-
-/**
- * :: Experimental ::
- * Summary of [[GeneralizedLinearRegression]] model and predictions.
- *
- * @param dataset Dataset to be summarized.
- * @param origModel Model to be summarized.  This is copied to create an internal
- *                  model which cannot be modified from outside.
- */
-@Since("2.0.0")
-@Experimental
-class GeneralizedLinearRegressionSummary private[regression] (
-    dataset: Dataset[_],
-    origModel: GeneralizedLinearRegressionModel) extends Serializable {
-
-  import GeneralizedLinearRegression._
-
-  /**
-   * Field in "predictions" which gives the predicted value of each instance.
-   * This is set to a new column name if the original model's `predictionCol` is not set.
-   */
-  @Since("2.0.0")
-  val predictionCol: String = {
-    if (origModel.isDefined(origModel.predictionCol) && origModel.getPredictionCol.nonEmpty) {
-      origModel.getPredictionCol
-    } else {
-      "prediction_" + java.util.UUID.randomUUID.toString
-    }
-  }
-
-  /**
-   * Private copy of model to ensure Params are not modified outside this class.
-   * Coefficients is not a deep copy, but that is acceptable.
-   *
-   * @note [[predictionCol]] must be set correctly before the value of [[model]] is set,
-   * and [[model]] must be set before [[predictions]] is set!
-   */
-  protected val model: GeneralizedLinearRegressionModel =
-    origModel.copy(ParamMap.empty).setPredictionCol(predictionCol)
-
-  /**
-   * Predictions output by the model's `transform` method.
-   */
-  @Since("2.0.0") @transient val predictions: DataFrame = model.transform(dataset)
-
-  private[regression] lazy val familyLink: FamilyAndLink = FamilyAndLink(model)
-
-  private[regression] lazy val family: Family = familyLink.family
-
-  private[regression] lazy val link: Link = familyLink.link
-
-  /** Number of instances in DataFrame predictions. */
-  @Since("2.2.0")
-  lazy val numInstances: Long = predictions.count()
-
-
-  /**
-   * Name of features. If the name cannot be retrieved from attributes,
-   * set default names to feature column name with numbered suffix "_0", "_1", and so on.
-   */
-  private[ml] lazy val featureNames: Array[String] = {
-    val featureAttrs = AttributeGroup.fromStructField(
-      dataset.schema(model.getFeaturesCol)).attributes
-    if (featureAttrs.isDefined) {
-      featureAttrs.get.map(_.name.get)
-    } else {
-      Array.tabulate[String](origModel.numFeatures)((x: Int) => model.getFeaturesCol + "_" + x)
-    }
-  }
-
-  /** The numeric rank of the fitted linear model. */
-  @Since("2.0.0")
-  lazy val rank: Long = if (model.getFitIntercept) {
-    model.coefficients.size + 1
-  } else {
-    model.coefficients.size
-  }
-
-  /** Degrees of freedom. */
-  @Since("2.0.0")
-  lazy val degreesOfFreedom: Long = numInstances - rank
-
-  /** The residual degrees of freedom. */
-  @Since("2.0.0")
-  lazy val residualDegreeOfFreedom: Long = degreesOfFreedom
-
-  /** The residual degrees of freedom for the null model. */
-  @Since("2.0.0")
-  lazy val residualDegreeOfFreedomNull: Long = {
-    if (model.getFitIntercept) numInstances - 1 else numInstances
-  }
-
-  private def label: Column = col(model.getLabelCol).cast(DoubleType)
-
-  private def prediction: Column = col(predictionCol)
-
-  private def weight: Column = {
-    if (!model.hasWeightCol) lit(1.0) else col(model.getWeightCol)
-  }
-
-  private def offset: Column = {
-    if (!model.hasOffsetCol) lit(0.0) else col(model.getOffsetCol).cast(DoubleType)
-  }
-
-  private[regression] lazy val devianceResiduals: DataFrame = {
-    val drUDF = udf { (y: Double, mu: Double, weight: Double) =>
-      val r = math.sqrt(math.max(family.deviance(y, mu, weight), 0.0))
-      if (y > mu) r else -1.0 * r
-    }
-    predictions.select(
-      drUDF(label, prediction, weight).as("devianceResiduals"))
-  }
-
-  private[regression] lazy val pearsonResiduals: DataFrame = {
-    val prUDF = udf { mu: Double => family.variance(mu) }
-    predictions.select(label.minus(prediction)
-      .multiply(sqrt(weight)).divide(sqrt(prUDF(prediction))).as("pearsonResiduals"))
-  }
-
-  private[regression] lazy val workingResiduals: DataFrame = {
-    val wrUDF = udf { (y: Double, mu: Double) => (y - mu) * link.deriv(mu) }
-    predictions.select(wrUDF(label, prediction).as("workingResiduals"))
-  }
-
-  private[regression] lazy val responseResiduals: DataFrame = {
-    predictions.select(label.minus(prediction).as("responseResiduals"))
-  }
-
-  /**
-   * Get the default residuals (deviance residuals) of the fitted model.
-   */
-  @Since("2.0.0")
-  def residuals(): DataFrame = devianceResiduals
-
-  /**
-   * Get the residuals of the fitted model by type.
-   *
-   * @param residualsType The type of residuals which should be returned.
-   *                      Supported options: deviance, pearson, working and response.
-   */
-  @Since("2.0.0")
-  def residuals(residualsType: String): DataFrame = {
-    residualsType match {
-      case "deviance" => devianceResiduals
-      case "pearson" => pearsonResiduals
-      case "working" => workingResiduals
-      case "response" => responseResiduals
-      case other => throw new UnsupportedOperationException(
-        s"The residuals type $other is not supported by Generalized Linear Regression.")
-    }
-  }
-
-  /**
-   * The deviance for the null model.
-   */
-  @Since("2.0.0")
-  lazy val nullDeviance: Double = {
-    val intercept: Double = if (!model.getFitIntercept) {
-      0.0
-    } else {
-      /*
-        Estimate intercept analytically when there is no offset, or when there is offset but
-        the model is Gaussian family with identity link. Otherwise, fit an intercept only model.
-       */
-      if (!model.hasOffsetCol ||
-        (model.hasOffsetCol && family == Gaussian && link == Identity)) {
-        val agg = predictions.agg(sum(weight.multiply(
-          label.minus(offset))), sum(weight)).first()
-        link.link(agg.getDouble(0) / agg.getDouble(1))
-      } else {
-        // Create empty feature column and fit intercept only model using param setting from model
-        val featureNull = "feature_" + java.util.UUID.randomUUID.toString
-        val paramMap = model.extractParamMap()
-        paramMap.put(model.featuresCol, featureNull)
-        if (family.name != "tweedie") {
-          paramMap.remove(model.variancePower)
-        }
-        val emptyVectorUDF = udf{ () => Vectors.zeros(0) }
-        model.parent.fit(
-          dataset.withColumn(featureNull, emptyVectorUDF()), paramMap
-        ).intercept
-      }
-    }
-    predictions.select(label, offset, weight).rdd.map {
-      case Row(y: Double, offset: Double, weight: Double) =>
-        family.deviance(y, link.unlink(intercept + offset), weight)
-    }.sum()
-  }
-
-  /**
-   * The deviance for the fitted model.
-   */
-  @Since("2.0.0")
-  lazy val deviance: Double = {
-    predictions.select(label, prediction, weight).rdd.map {
-      case Row(label: Double, pred: Double, weight: Double) =>
-        family.deviance(label, pred, weight)
-    }.sum()
-  }
-
-  /**
-   * The dispersion of the fitted model.
-   * It is taken as 1.0 for the "binomial" and "poisson" families, and otherwise
-   * estimated by the residual Pearson's Chi-Squared statistic (which is defined as
-   * sum of the squares of the Pearson residuals) divided by the residual degrees of freedom.
-   */
-  @Since("2.0.0")
-  lazy val dispersion: Double = if (
-    model.getFamily.toLowerCase(Locale.ROOT) == Binomial.name ||
-      model.getFamily.toLowerCase(Locale.ROOT) == Poisson.name) {
-    1.0
-  } else {
-    val rss = pearsonResiduals.agg(sum(pow(col("pearsonResiduals"), 2.0))).first().getDouble(0)
-    rss / degreesOfFreedom
-  }
-
-  /** Akaike Information Criterion (AIC) for the fitted model. */
-  @Since("2.0.0")
-  lazy val aic: Double = {
-    val weightSum = predictions.select(weight).agg(sum(weight)).first().getDouble(0)
-    val t = predictions.select(
-      label, prediction, weight).rdd.map {
-        case Row(label: Double, pred: Double, weight: Double) =>
-          (label, pred, weight)
-    }
-    family.aic(t, deviance, numInstances, weightSum) + 2 * rank
-  }
-}
-
-/**
- * :: Experimental ::
- * Summary of [[GeneralizedLinearRegression]] fitting and model.
- *
- * @param dataset Dataset to be summarized.
- * @param origModel Model to be summarized.  This is copied to create an internal
- *                  model which cannot be modified from outside.
- * @param diagInvAtWA diagonal of matrix (A^T * W * A)^-1 in the last iteration
- * @param numIterations number of iterations
- * @param solver the solver algorithm used for model training
- */
-@Since("2.0.0")
-@Experimental
-class GeneralizedLinearRegressionTrainingSummary private[regression] (
-    dataset: Dataset[_],
-    origModel: GeneralizedLinearRegressionModel,
-    private val diagInvAtWA: Array[Double],
-    @Since("2.0.0") val numIterations: Int,
-    @Since("2.0.0") val solver: String)
-  extends GeneralizedLinearRegressionSummary(dataset, origModel) with Serializable {
-
-  import GeneralizedLinearRegression._
-
-  /**
-   * Whether the underlying `WeightedLeastSquares` using the "normal" solver.
-   */
-  private[ml] val isNormalSolver: Boolean = {
-    diagInvAtWA.length != 1 || diagInvAtWA(0) != 0
-  }
-
-  /**
-   * Standard error of estimated coefficients and intercept.
-   * This value is only available when the underlying `WeightedLeastSquares`
-   * using the "normal" solver.
-   *
-   * If `GeneralizedLinearRegression.fitIntercept` is set to true,
-   * then the last element returned corresponds to the intercept.
-   */
-  @Since("2.0.0")
-  lazy val coefficientStandardErrors: Array[Double] = {
-    if (isNormalSolver) {
-      diagInvAtWA.map(_ * dispersion).map(math.sqrt)
-    } else {
-      throw new UnsupportedOperationException(
-        "No Std. Error of coefficients available for this GeneralizedLinearRegressionModel")
-    }
-  }
-
-  /**
-   * T-statistic of estimated coefficients and intercept.
-   * This value is only available when the underlying `WeightedLeastSquares`
-   * using the "normal" solver.
-   *
-   * If `GeneralizedLinearRegression.fitIntercept` is set to true,
-   * then the last element returned corresponds to the intercept.
-   */
-  @Since("2.0.0")
-  lazy val tValues: Array[Double] = {
-    if (isNormalSolver) {
-      val estimate = if (model.getFitIntercept) {
-        Array.concat(model.coefficients.toArray, Array(model.intercept))
-      } else {
-        model.coefficients.toArray
-      }
-      estimate.zip(coefficientStandardErrors).map { x => x._1 / x._2 }
-    } else {
-      throw new UnsupportedOperationException(
-        "No t-statistic available for this GeneralizedLinearRegressionModel")
-    }
-  }
-
-  /**
-   * Two-sided p-value of estimated coefficients and intercept.
-   * This value is only available when the underlying `WeightedLeastSquares`
-   * using the "normal" solver.
-   *
-   * If `GeneralizedLinearRegression.fitIntercept` is set to true,
-   * then the last element returned corresponds to the intercept.
-   */
-  @Since("2.0.0")
-  lazy val pValues: Array[Double] = {
-    if (isNormalSolver) {
-      if (model.getFamily.toLowerCase(Locale.ROOT) == Binomial.name ||
-        model.getFamily.toLowerCase(Locale.ROOT) == Poisson.name) {
-        tValues.map { x => 2.0 * (1.0 - dist.Gaussian(0.0, 1.0).cdf(math.abs(x))) }
-      } else {
-        tValues.map { x =>
-          2.0 * (1.0 - dist.StudentsT(degreesOfFreedom.toDouble).cdf(math.abs(x)))
-        }
-      }
-    } else {
-      throw new UnsupportedOperationException(
-        "No p-value available for this GeneralizedLinearRegressionModel")
-    }
-  }
-
-  /**
-   * Coefficients with statistics: feature name, coefficients, standard error, tValue and pValue.
-   */
-  private[ml] lazy val coefficientsWithStatistics: Array[
-    (String, Double, Double, Double, Double)] = {
-    var featureNamesLocal = featureNames
-    var coefficientsArray = model.coefficients.toArray
-    var index = Array.range(0, coefficientsArray.length)
-    if (model.getFitIntercept) {
-      featureNamesLocal = featureNamesLocal :+ "(Intercept)"
-      coefficientsArray = coefficientsArray :+ model.intercept
-      // Reorder so that intercept comes first
-      index = (coefficientsArray.length - 1) +: index
-    }
-    index.map { i =>
-      (featureNamesLocal(i), coefficientsArray(i), coefficientStandardErrors(i),
-        tValues(i), pValues(i))
-    }
-  }
-
-  override def toString: String = {
-    if (isNormalSolver) {
-
-      def round(x: Double): String = {
-        BigDecimal(x).setScale(4, BigDecimal.RoundingMode.HALF_UP).toString
-      }
-
-      val colNames = Array("Feature", "Estimate", "Std Error", "T Value", "P Value")
-
-      val data = coefficientsWithStatistics.map { row =>
-        val strRow = row.productIterator.map { cell =>
-          val str = cell match {
-            case s: String => s
-            case n: Double => round(n)
-          }
-          // Truncate if length > 20
-          if (str.length > 20) {
-            str.substring(0, 17) + "..."
-          } else {
-            str
-          }
-        }
-        strRow.toArray
-      }
-
-      // Compute the width of each column
-      val colWidths = colNames.map(_.length)
-      data.foreach { strRow =>
-        strRow.zipWithIndex.foreach { case (cell: String, i: Int) =>
-          colWidths(i) = math.max(colWidths(i), cell.length)
-        }
-      }
-
-      val sb = new StringBuilder
-
-      // Output coefficients with statistics
-      sb.append("Coefficients:\n")
-      colNames.zipWithIndex.map { case (colName: String, i: Int) =>
-        StringUtils.leftPad(colName, colWidths(i))
-      }.addString(sb, "", " ", "\n")
-
-      data.foreach { case strRow: Array[String] =>
-        strRow.zipWithIndex.map { case (cell: String, i: Int) =>
-          StringUtils.leftPad(cell.toString, colWidths(i))
-        }.addString(sb, "", " ", "\n")
-      }
-
-      sb.append("\n")
-      sb.append(s"(Dispersion parameter for ${family.name} family taken to be " +
-        s"${round(dispersion)})")
-
-      sb.append("\n")
-      val nd = s"Null deviance: ${round(nullDeviance)} on $degreesOfFreedom degrees of freedom"
-      val rd = s"Residual deviance: ${round(deviance)} on $residualDegreeOfFreedom degrees of " +
-        "freedom"
-      val l = math.max(nd.length, rd.length)
-      sb.append(StringUtils.leftPad(nd, l))
-      sb.append("\n")
-      sb.append(StringUtils.leftPad(rd, l))
-
-      if (family.name != "tweedie") {
-        sb.append("\n")
-        sb.append(s"AIC: " + round(aic))
-      }
-
-      sb.toString()
-    } else {
-      throw new UnsupportedOperationException(
-        "No summary available for this GeneralizedLinearRegressionModel")
     }
   }
 }
