+++ mllib/src/test/scala/org/apache/spark/ml/clustering/KMeansSuite.scala
@@ -18,10 +18,10 @@
 package org.apache.spark.ml.clustering
 
 import scala.util.Random
-
 import org.apache.spark.SparkFunSuite
 import org.apache.spark.ml.linalg.{Vector, Vectors}
 import org.apache.spark.ml.param.ParamMap
+import org.apache.spark.ml.summary.KMeansSummary
 import org.apache.spark.ml.util.{DefaultReadWriteTest, MLTestingUtils}
 import org.apache.spark.mllib.clustering.{KMeans => MLlibKMeans}
 import org.apache.spark.mllib.util.MLlibTestSparkContext
